refactor: extract toPrimitive and toProeprtyKey to single helpers
