+++ packages/babel-helpers/src/helpers/toPrimitive.ts
@@ -1,0 +1,16 @@
+/* @minVersion 7.1.5 */
+
+// https://tc39.es/ecma262/#sec-toprimitive
+export default function toPrimitive(
+  input: unknown,
+  hint: "default" | "string" | "number" | void,
+) {
+  if (typeof input !== "object" || input === null) return input;
+  var prim = input[Symbol.toPrimitive];
+  if (prim !== undefined) {
+    var res = prim.call(input, hint || "default");
+    if (typeof res !== "object") return res;
+    throw new TypeError("@@toPrimitive must return a primitive value.");
+  }
+  return (hint === "string" ? String : Number)(input);
+}
